regen: more renaming and moving around
