+++ src/Regen.Core/Compiler/Digest/DigestParser.cs
@@ -1,15 +1,14 @@
-﻿using System;
-using System.Linq;
+﻿using System.Linq;
 using System.Text.RegularExpressions;
 using Regen.Helpers;
 
-namespace Regen.Compiler {
+namespace Regen.Compiler.Digest {
 
     /// <summary>
     ///     A parser finds #if <see cref="DefineMarker"/>, passes contents to an <see cref="Interpreter"/> and then places the output inside the #else block.
     /// </summary>
     /// <remarks>If you are looking to just parse a template without regions and so on, please refer to <see cref="Interpreter"/>.<see cref="Interpreter.Interpret(System.Collections.Generic.Dictionary{string,Regen.DataTypes.Data})"/></remarks>
-    public class Parser {
+    public class DigestParser {
         /// <summary>
         ///     The regex options used to in parsing.
         /// </summary>
